learningplatform_nowy2/backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=11.3.0
python-dotenv==1.0.1
firebase-admin==6.4.0
djangorestframework-simplejwt==5.3.1
gunicorn
PyJWT>=2.5.0
sympy==1.12